requirements.txt
streamlit>=1.30.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
bcrypt>=4.0.0
pandas>=2.0.0
plotly>=5.18.0
python-dotenv>=1.0.0
requests>=2.31.0
python-dateutil>=2.8.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
